+++ MaXrd/Kernel/init.m
@@ -2,8 +2,8 @@
 
 
 (*---* Load definitions *---*)
-Get["MaXrd`Kernel`Definitions`"]
+Get["MaXrd`Kernel`Definitions`"];
 
 (*---* Check version *---*)
 If[$Notebooks && $VersionNumber < 11.3,
-Print["MaXrd " <> $MaXrdVersion <> " requires Mathematica version 11.3 or higher."]]
+Print["MaXrd " <> $MaXrdVersion <> " requires Mathematica version 11.3 or higher."]];